+++ backend/__init__.py
@@ -1,3 +1,3 @@
 """Package level information"""
 
-__version__ = "7.5.2"
+__version__ = "7.6.1"